Refactor in-proc eventing code for re-use (#4866)
