+++ src/Microsoft.Diagnostics.Monitoring.StartupHook/Exceptions/Eventing/ExceptionsEventSource.cs
@@ -1,11 +1,9 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using Microsoft.Diagnostics.Monitoring.StartupHook.Eventing;
 using System;
 using System.Diagnostics.Tracing;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-using System.Threading;
 
 namespace Microsoft.Diagnostics.Monitoring.StartupHook.Exceptions.Eventing
 {
@@ -13,39 +11,8 @@ namespace Microsoft.Diagnostics.Monitoring.StartupHook.Exceptions.Eventing
     // likely be sending many events every second. Avoid any APIs that use params/varargs
     // style calls and avoid heap allocations as much as possible.
     [EventSource(Name = ExceptionEvents.SourceName)]
-    internal sealed class ExceptionsEventSource : EventSource
+    internal sealed class ExceptionsEventSource : AbstractMonitorEventSource
     {
-        // Arrays are expected to have a 16-bit field for the length of the array.
-        // The length of the array is the number of elements, not the number of bytes.
-        private const int ArrayLengthFieldSize = sizeof(short);
-
-        /// <summary>
-        /// Amount of time to wait before sending batches of event source events in order to
-        /// avoid real-time buffering issues in the runtime eventing infrastructure and the
-        /// trace event library event processor.
-        /// </summary>
-        /// <remarks>
-        /// See: https://github.com/dotnet/runtime/issues/76704
-        /// </remarks>
-        private static readonly TimeSpan EventSourceBufferAvoidanceTimeout = TimeSpan.FromMilliseconds(200);
-
-        private readonly Timer _flushEventsTimer;
-
-        // NOTE: Arrays with a non-"byte" element type are not supported well by in-proc EventListener
-        // when using self-describing event format. This format is used to easily support event pipe listening,
-        // which is the primary use of this event source.
-        public ExceptionsEventSource()
-            : base(EventSourceSettings.EtwSelfDescribingEventFormat)
-        {
-            _flushEventsTimer = new Timer(FlushTimerTick);
-        }
-
-        protected override void Dispose(bool disposing)
-        {
-            _flushEventsTimer.Dispose();
-            base.Dispose(disposing);
-        }
-
         [Event(ExceptionEvents.EventIds.ExceptionGroup)]
         public void ExceptionGroup(
             ulong ExceptionGroupId,
@@ -60,9 +27,7 @@ namespace Microsoft.Diagnostics.Monitoring.StartupHook.Exceptions.Eventing
             SetValue(ref data[ExceptionEvents.ExceptionGroupPayloads.ThrowingMethodId], ThrowingMethodId);
             SetValue(ref data[ExceptionEvents.ExceptionGroupPayloads.ILOffset], ILOffset);
 
-            WriteEventCore(ExceptionEvents.EventIds.ExceptionGroup, data);
-
-            RestartFlushingEventTimer();
+            WriteEventWithFlushing(ExceptionEvents.EventIds.ExceptionGroup, data);
         }
 
         [Event(ExceptionEvents.EventIds.ExceptionInstance)]
@@ -88,9 +53,7 @@ namespace Microsoft.Diagnostics.Monitoring.StartupHook.Exceptions.Eventing
             SetValue(ref data[ExceptionEvents.ExceptionInstancePayloads.Timestamp], Timestamp.ToFileTimeUtc());
             SetValue(ref data[ExceptionEvents.ExceptionInstancePayloads.InnerExceptionIds], innerExceptionIdsSpan);
 
-            WriteEventCore(ExceptionEvents.EventIds.ExceptionInstance, data);
-
-            RestartFlushingEventTimer();
+            WriteEventWithFlushing(ExceptionEvents.EventIds.ExceptionInstance, data);
         }
 
         [Event(ExceptionEvents.EventIds.ClassDescription)]
@@ -111,9 +74,7 @@ namespace Microsoft.Diagnostics.Monitoring.StartupHook.Exceptions.Eventing
             SetValue(ref data[NameIdentificationEvents.ClassDescPayloads.Flags], Flags);
             SetValue(ref data[NameIdentificationEvents.ClassDescPayloads.TypeArgs], typeArgsSpan);
 
-            WriteEventCore(ExceptionEvents.EventIds.ClassDescription, data);
-
-            RestartFlushingEventTimer();
+            WriteEventWithFlushing(ExceptionEvents.EventIds.ClassDescription, data);
         }
 
         [Event(ExceptionEvents.EventIds.FunctionDescription)]
@@ -141,9 +102,7 @@ namespace Microsoft.Diagnostics.Monitoring.StartupHook.Exceptions.Eventing
             SetValue(ref data[NameIdentificationEvents.FunctionDescPayloads.TypeArgs], typeArgsSpan);
             SetValue(ref data[NameIdentificationEvents.FunctionDescPayloads.ParameterTypes], parameterTypesSpan);
 
-            WriteEventCore(ExceptionEvents.EventIds.FunctionDescription, data);
-
-            RestartFlushingEventTimer();
+            WriteEventWithFlushing(ExceptionEvents.EventIds.FunctionDescription, data);
         }
 
         [Event(ExceptionEvents.EventIds.ModuleDescription)]
@@ -157,9 +116,7 @@ namespace Microsoft.Diagnostics.Monitoring.StartupHook.Exceptions.Eventing
             SetValue(ref data[NameIdentificationEvents.ModuleDescPayloads.ModuleId], ModuleId);
             SetValue(ref data[NameIdentificationEvents.ModuleDescPayloads.Name], namePinned);
 
-            WriteEventCore(ExceptionEvents.EventIds.ModuleDescription, data);
-
-            RestartFlushingEventTimer();
+            WriteEventWithFlushing(ExceptionEvents.EventIds.ModuleDescription, data);
         }
 
         [Event(ExceptionEvents.EventIds.StackFrameDescription)]
@@ -174,9 +131,7 @@ namespace Microsoft.Diagnostics.Monitoring.StartupHook.Exceptions.Eventing
             SetValue(ref data[ExceptionEvents.StackFrameIdentifierPayloads.FunctionId], FunctionId);
             SetValue(ref data[ExceptionEvents.StackFrameIdentifierPayloads.ILOffset], ILOffset);
 
-            WriteEventCore(ExceptionEvents.EventIds.StackFrameDescription, data);
-
-            RestartFlushingEventTimer();
+            WriteEventWithFlushing(ExceptionEvents.EventIds.StackFrameDescription, data);
         }
 
         [Event(ExceptionEvents.EventIds.TokenDescription)]
@@ -194,127 +149,13 @@ namespace Microsoft.Diagnostics.Monitoring.StartupHook.Exceptions.Eventing
             SetValue(ref data[NameIdentificationEvents.TokenDescPayloads.OuterToken], OuterToken);
             SetValue(ref data[NameIdentificationEvents.TokenDescPayloads.Name], namePinned);
 
-            WriteEventCore(ExceptionEvents.EventIds.TokenDescription, data);
-
-            RestartFlushingEventTimer();
+            WriteEventWithFlushing(ExceptionEvents.EventIds.TokenDescription, data);
         }
 
         [Event(ExceptionEvents.EventIds.Flush)]
-        private void Flush()
+        protected override void Flush()
         {
             WriteEvent(ExceptionEvents.EventIds.Flush);
-        }
-
-        [NonEvent]
-        private void FlushTimerTick(object? state)
-        {
-            Flush();
-        }
-
-        [NonEvent]
-        private void RestartFlushingEventTimer()
-        {
-            // This will reset the timer to fire after the specified time period. If the timer is already
-            // started, it will be reset. If it already finished, it will be started again.
-            _flushEventsTimer.Change(EventSourceBufferAvoidanceTimeout, Timeout.InfiniteTimeSpan);
-        }
-
-        [NonEvent]
-        private unsafe void WriteEventCore(int eventId, Span<EventData> data)
-        {
-            fixed (EventData* dataPtr = data)
-            {
-                WriteEventCore(eventId, data.Length, dataPtr);
-            }
-        }
-
-        [NonEvent]
-        private static unsafe void SetValue<T>(ref EventData data, in T value) where T : unmanaged
-        {
-            data.DataPointer = (nint)Unsafe.AsPointer(ref Unsafe.AsRef(value));
-            data.Size = sizeof(T);
-        }
-
-        [NonEvent]
-        private static unsafe void SetValue(ref EventData data, in Span<byte> value)
-        {
-            // It is expected that the Span is a wrapper around an array-like value
-            // that is pinned or is unmovable from the perspective of GC (e.g. stackalloc).
-            // Otherwise, GC might relocate the value after its address is acquired and
-            // potentially cause access violations or misinterpretation of the data.
-            data.DataPointer = (nint)Unsafe.AsPointer(ref value.GetPinnableReference());
-            data.Size = value.Length;
-        }
-
-        [NonEvent]
-        private static void SetValue(ref EventData data, in PinnedData value)
-        {
-            data.DataPointer = value.Address;
-            data.Size = value.Size;
-        }
-
-        private static int GetArrayDataSize<T>(T[] data) where T : unmanaged
-        {
-            // Arrays are written with a length prefix + the data as bytes
-            return ArrayLengthFieldSize + Unsafe.SizeOf<T>() * data.Length;
-        }
-
-        private static unsafe void FillArrayData<T>(Span<byte> target, T[] source) where T : unmanaged
-        {
-#if DEBUG
-            // Double-check that the Span is correctly sized. This shouldn't be encountered
-            // at runtime so long as Span is constructed correctly using GetArrayDataSize.
-            if (target.Length != GetArrayDataSize(source))
-                throw new ArgumentOutOfRangeException(nameof(source));
-#endif
-
-            // First, copy the length of the array to the beginning of the data
-            short length = checked((short)source.Length);
-            ReadOnlySpan<byte> lengthSpan = new(Unsafe.AsPointer(ref Unsafe.AsRef(length)), ArrayLengthFieldSize);
-            lengthSpan.CopyTo(target);
-
-            // Second, copy the array data as bytes
-            if (source.Length > 0)
-            {
-                ReadOnlySpan<byte> dataSpan = MemoryMarshal.CreateReadOnlySpan(
-                    ref Unsafe.As<T, byte>(ref MemoryMarshal.GetArrayDataReference(source)),
-                    Unsafe.SizeOf<T>() * source.Length);
-
-                dataSpan.CopyTo(target.Slice(ArrayLengthFieldSize));
-            }
-        }
-
-        private struct PinnedData : IDisposable
-        {
-            private readonly GCHandle _handle;
-
-            private PinnedData(scoped in object? value, int Size)
-            {
-                _handle = GCHandle.Alloc(value, GCHandleType.Pinned);
-
-                this.Size = Size;
-            }
-
-            public static PinnedData Create(in string? value)
-            {
-                if (null == value)
-                {
-                    return new PinnedData(null, 0);
-                }
-                return new PinnedData(value, sizeof(char) * (value.Length + 1));
-            }
-
-            public void Dispose()
-            {
-                if (_handle.IsAllocated)
-                {
-                    _handle.Free();
-                }
-            }
-
-            public IntPtr Address => _handle.AddrOfPinnedObject();
-
-            public int Size { get; }
         }
     }
 }
